+++ src/lys/apps/file_management/modules/file_import/services.py
@@ -522,3 +522,4 @@ class AbstractImportService(abc.ABC):
                     report=report,
                 )
                 session.commit()
+                raise